pyproject.toml
[tool.poetry]
name = "python-django-team"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "market"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
ipython = "^8.14.0"
django-extensions = "^3.2.3"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.2"
django-jinja = "^2.11.0"
pillow = "^10.1.0"
django-debug-toolbar = "^4.2.0"
django-filter = "^23.4"
django-redis = "^5.4.0"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
pydantic = "^2.5.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 119
